refactor: keep custom element wiring under the source length gate
