+++ crates/vize_atelier_sfc/src/compile/tests.rs
@@ -1,3 +1,4 @@
+mod custom_elements;
 mod define_props_regressions;
 mod options_api_events;
 mod vapor;
@@ -2921,84 +2922,6 @@ const visible = true
     insta::assert_snapshot!(result.code.as_str());
 }
 
-#[test]
-fn test_script_setup_sfc_custom_elements_preserve_imports_and_pascal_case_intrinsics() {
-    let source = r#"<script setup lang="ts">
-import { TresCanvas } from '@tresjs/core'
-const visible = true
-</script>
-
-<template>
-  <TresCanvas>
-    <TresMesh v-if="visible">
-      <TresSpotLight></TresSpotLight>
-    </TresMesh>
-  </TresCanvas>
-</template>"#;
-
-    let descriptor = parse_sfc(source, SfcParseOptions::default()).expect("Failed to parse SFC");
-    let opts = SfcCompileOptions {
-        script: ScriptCompileOptions {
-            is_ts: true,
-            ..Default::default()
-        },
-        template: TemplateCompileOptions {
-            is_ts: true,
-            custom_renderer: true,
-            custom_elements: vec!["Tres*".into()],
-            ..Default::default()
-        },
-        ..Default::default()
-    };
-    let result = compile_sfc(&descriptor, opts).expect("Failed to compile SFC");
-
-    assert!(
-        result.code.contains(r#"_createBlock(_unref(TresCanvas)"#),
-        "{}",
-        result.code
-    );
-    assert!(
-        result.code.contains(r#"_createElementBlock("TresMesh""#),
-        "{}",
-        result.code
-    );
-    assert!(
-        result
-            .code
-            .contains(r#"_createElementVNode("TresSpotLight""#),
-        "{}",
-        result.code
-    );
-    assert!(!result.code.contains(r#"_resolveComponent("TresCanvas")"#));
-    assert!(!result.code.contains(r#"_resolveComponent("TresMesh")"#));
-    assert!(
-        !result
-            .code
-            .contains(r#"_resolveComponent("TresSpotLight")"#)
-    );
-}
-
-#[test]
-fn test_normal_script_sfc_vapor_output_mode() {
-    let source = r#"<script>
-export default {
-  name: 'NormalVapor'
-}
-</script>
-
-<template>
-  <div>Hello</div>
-</template>"#;
-
-    let descriptor = parse_sfc(source, SfcParseOptions::default()).expect("Failed to parse SFC");
-    let opts = SfcCompileOptions {
-        vapor: true,
-        ..Default::default()
-    };
-    let result = compile_sfc(&descriptor, opts).expect("Failed to compile SFC");
-
-    insta::assert_snapshot!(result.code.as_str());
-}
 /// A class component's `@Component({ components: { Foo } })` registration and
 /// its members must thread through compile output exactly like the equivalent
 /// Options API component: `<Foo/>` resolves via `_resolveComponent("Foo")`
