+++ npm/builder/vite/src/test.ts
@@ -2,6 +2,7 @@ import "./hmr.test.ts";
 import "./native-request-gates.test.ts";
 import "./compiler.test.ts";
 import "./compiler-src-imports.test.ts";
+import "./compiler-custom-elements.test.ts";
 import "./compile-options.test.ts";
 import "./config.test.ts";
 import "./internal/config-bridge.test.ts";
